Added trials, settings, and phases
The application now runs in 3 phases. Phase 0 resets the environment and
waits for all objects to be clicked (only click currently). Phase 1
lines the objects up and gives the user 25 seconds to rearrange them
(they currently can't be moved with mouse... gotta make that work next).
Phase 2 waits for 10 seconds and then returns to phase 0.

Each phase has a 1 second transition between them, but this can be
eliminated if needed.

The mouse click and drag is the last main functionality to be added.
I've tested everything but I'm sure there are still a few bugs (for
instance, the hit boxes on some of the shapes are a little... rough).