Logger Summary File
Added code to allow for a summary file of start and end conditions for
each trial to be saved along with the primary log files. The logger will
need to still be updated to utilitze the subject identifier but should
be otherwise complete. I haven't done full testing on it yet though so
keep an eye out for bugs!