Centered Blobs
Adjusted all the blobs to be centered around their center point, instead
of off to the side as they were before. This should fix a lot of the
grabbing problems. The fix involved creating a parent game object, so I
still have to modify the script to work on child game objects.