Floksy Assets, Leap Move Task
Imported all the assets Floksy sent me. They look really nice. Also
polished up the move task. Now all it's missing is the logger!

Deleted unneeded files

There was a massively huge texture that was keeping it from pushing the
commit. Deleted it, since we're not using it.

Fixed Logger and Leap Move Task

Added Audio and Video

Clock ticks now, and the tv plays a video on start (reset it with tab).

Removed Large Textures